requirements.txt
langchain>=0.2.0
langchain-community>=0.0.30
langchain-core>=0.1.50
faiss-cpu
transformers>=4.41.2
sentence-transformers
streamlit
python-dotenv
torch